split data-types
