+++ src/components/content/entity.tsx
@@ -17,7 +17,8 @@ import { CourseNavigation } from '@/components/navigation/course-navigation'
 import { ShareModalProps } from '@/components/user-tools/share-modal'
 import { UserTools } from '@/components/user-tools/user-tools'
 import { useInstanceData } from '@/contexts/instance-context'
-import { EntityData, FrontendContentNode } from '@/data-types'
+import { EntityData } from '@/data-types'
+import { FrontendContentNode } from '@/frontend-node-types'
 import { getIconByTypename } from '@/helper/icon-by-entity-type'
 import { replacePlaceholders } from '@/helper/replace-placeholders'
 import { getHistoryUrl } from '@/helper/urls/get-history-url'
