+++ src/components/content/exercises/sc-mc-exercise.tsx
@@ -9,7 +9,7 @@ import { Feedback } from './feedback'
 import { FaIcon } from '@/components/fa-icon'
 import { isPrintMode } from '@/components/print-mode'
 import { useInstanceData } from '@/contexts/instance-context'
-import { EdtrPluginScMcExercise } from '@/data-types'
+import { EdtrPluginScMcExercise } from '@/frontend-node-types'
 import { hasVisibleContent } from '@/helper/has-visible-content'
 import { NodePath, RenderNestedFunction } from '@/schema/article-renderer'
