+++ src/components/content/injection.tsx
@@ -2,7 +2,8 @@ import { useState, useEffect } from 'react'
 
 import { LoadingSpinner } from '../loading/loading-spinner'
 import { useInstanceData } from '@/contexts/instance-context'
-import { FrontendContentNode, SlugProps } from '@/data-types'
+import { SlugProps } from '@/data-types'
+import { FrontendContentNode } from '@/frontend-node-types'
 import type { RenderNestedFunction } from '@/schema/article-renderer'
 
 export interface InjectionProps {
