+++ src/edtr-io/plugins/page-team/frontend.tsx
@@ -1,5 +1,5 @@
 import { PageTeamRenderer } from './renderer'
-import type { FrontendPageTeamNode } from '@/data-types'
+import type { FrontendPageTeamNode } from '@/frontend-node-types'
 
 export function PageTeamAdapter({ data }: FrontendPageTeamNode) {
   return <PageTeamRenderer data={data} />
