+++ src/fetcher/convert-state.ts
@@ -1,5 +1,5 @@
 import { render } from '../../external/legacy_render'
-import { FrontendContentNode } from '@/data-types'
+import { FrontendContentNode } from '@/frontend-node-types'
 import { convert } from '@/schema/convert-edtr-io-state'
 import { convertLegacyState } from '@/schema/convert-legacy-state'
