+++ src/fetcher/create-exercises.ts
@@ -9,7 +9,7 @@ import {
   SolutionEdtrState,
   FrontendExerciseGroupNode,
   FrontendSolutionNode,
-} from '@/data-types'
+} from '@/frontend-node-types'
 import { hasVisibleContent } from '@/helper/has-visible-content'
 import { shuffleArray } from '@/helper/shuffle-array'
 import { convert, ConvertNode } from '@/schema/convert-edtr-io-state'
