+++ src/fetcher/prettify-links.tsx
@@ -2,7 +2,8 @@ import { request } from 'graphql-request'
 
 import { idsQuery } from './query-ids'
 import { endpoint } from '@/api/endpoint'
-import { RequestPageData, FrontendContentNode } from '@/data-types'
+import { RequestPageData } from '@/data-types'
+import { FrontendContentNode } from '@/frontend-node-types'
 import { hasSpecialUrlChars } from '@/helper/urls/check-special-url-chars'
 
 export async function prettifyLinks(pageData: RequestPageData) {
